Moved page files to their proper folders. issue [#89]
