+++ app/src/main/java/com/example/bigowlapp/Services/ViewAnotherUserActivity.java
@@ -1,14 +1,14 @@
-package com.example.bigowlapp;
+package com.example.bigowlapp.Services;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.annotation.SuppressLint;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
 
+import com.example.bigowlapp.R;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
