+++ app/src/main/res/layout/activity_invitation_confirmation.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".InvitationConfirmationActivity"
+    tools:context=".Services.InvitationConfirmationActivity"
     android:orientation="horizontal">
 
     <RelativeLayout
