+++ app/src/main/res/layout/activity_monitoring_group_page.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MonitoringGroupPageActivity">
+    tools:context=".ActivityPage.MonitoringGroupPageActivity">
 
     <TextView
         android:id="@+id/textView_groupName"
