+++ app/src/main/res/layout/activity_send_sms_invitation.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".SendSmsInvitationActivity">
+    tools:context=".Services.SendSmsInvitationActivity">
 
 
     <EditText
