+++ app/src/main/res/layout/activity_supervised_group_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SupervisedGroupListActivity">
+    tools:context=".Services.SupervisedGroupListActivity">
 
     <TextView
         android:id="@+id/pageTitle"
