+++ app/src/main/res/layout/activity_supervised_group_page.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SupervisedGroupPageActivity">
+    tools:context=".ActivityPage.SupervisedGroupPageActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
