+++ app/src/main/res/layout/activity_view_another_user.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ViewAnotherUserActivity">
+    tools:context=".Services.ViewAnotherUserActivity">
 
     <Button
         android:id="@+id/SupRequest"
